move locking code to single file (WP-808)
